fix(version): properly link build time info
