+++ Makefile
@@ -2,12 +2,12 @@
 VERSION?=v0.2.0
 IMG?=fintechstudios/ververica-platform-k8s-controller
 PKG=github.com/fintechstudios.com/ververica-platform-k8s-controller
-VERSION_PKG=$(PKG)/controllers/version/version
+VERSION_PKG=main
 BUILD=$(shell date -u +'%Y-%m-%dT%H:%M:%SZ')
 # Produce CRDs that work back to Kubernetes 1.11 (no version conversion)
 CRD_OPTIONS?="crd:trivialVersions=true"
 
-LD_FLAGS="-X $(VERSION_PKG).controllerVersion=$(VERSION) -X $(VERSION_PKG).gitCommit=$(GIT_COMMIT) -X $(VERSION_PKG).buildDate=$(BUILD)"
+LD_FLAGS="-X $(VERSION_PKG).controllerVersion='$(VERSION)' -X $(VERSION_PKG).gitCommit='$(GIT_COMMIT)' -X $(VERSION_PKG).buildDate='$(BUILD)'"
 
 TEST_CLUSTER_NAME=ververica-platform-k8s-controller-cluster
